Many things
create the adaptive guis
Add motivationnal quotes.
Keep implementing single time to probe.
