+++ gui_pulser.py
@@ -24,13 +24,8 @@ import gui_signal_generator
 
 import time
 
-# For adaptive protocols
-import protocol_model
-import protocol_bayes
-
 
 # Debug stuff.
-_fc._debug_enabled = False
 _debug_enabled     = False
 
 def _debug(*a):
@@ -52,6 +47,7 @@ class GuiMainPulseSequence(egg.gui.Window):
             The session of the fpga must already be open.    
         """    
         _debug('GuiMainPulseSequence:__init__')
+        _debug('Don’t watch the clock; do what it does. Keep going. – Sam Levenson')
         
         # Run the basic stuff for the initialization
         egg.gui.Window.__init__(self, title=name, size=size)
@@ -828,6 +824,7 @@ class GUIPulseBuilder(egg.gui.Window):
         Initialize 
         """    
         _debug('GUIPulseBuilder:__init__')
+        _debug('Be a fruitloop in a world of Cheerios.')
         
         # Run the basic stuff for the initialization
         egg.gui.Window.__init__(self, title=name, size=size)
@@ -1072,6 +1069,7 @@ class GUIPredefined(egg.gui.Window):
 
         """    
         _debug('GUIPredefined:__init__')
+        _debug('The best way to predict your future is to create it. – Abraham Lincoln')
         
         # Run the basic stuff for the initialization
         egg.gui.Window.__init__(self, title=name, size=size)
@@ -3945,76 +3943,39 @@ class GUIT1probeOneTime(egg.gui.Window):
         self.connect(self.button_prepare_experiment.signal_clicked, 
                      self.button_prepare_experiment_clicked)
             
+        # A button for preparing stuff
+        self.button_prepare_experiment = egg.gui.Button('Prepare',
+                                                        tip='Prepare the measurement before running')
+        self.place_object(self.button_prepare_experiment, row=1, column=0)
+        self.connect(self.button_prepare_experiment.signal_clicked, 
+                     self.button_prepare_experiment_clicked)
+            
         # tree dictionnarry for the settings
-        self.treeDic_settings = egg.gui.TreeDictionary(autosettings_path='setting_adaptive_T1Bayes')
+        self.treeDic_settings = egg.gui.TreeDictionary(autosettings_path='setting_T1_singleTime')
         self.place_object(self.treeDic_settings, row=2, column=0)
 
-        self.treeDic_settings.add_parameter('Power_ms_+1', -20, 
+        self.treeDic_settings.add_parameter('Power', -20, 
                                             type='float', step=1, 
                                             bounds=[-50,30], suffix=' dBm',
-                                            tip='Constant power of the pi pulse of ms=+1')
-        self.treeDic_settings.add_parameter('Frequency_ms_+1', 3, 
+                                            tip='Constant power of the RF')
+        self.treeDic_settings.add_parameter('Frequency', 3, 
                                             type='float', step=0.1, 
                                             bounds=[0,10], suffix=' GHz',
-                                            tip='Frequency of the pi pulse of ms=+1')     
-        self.treeDic_settings.add_parameter('dt_pi_pulse_ms_+1', 0.3, 
-                                            type='float', step=0.1, 
-                                            bounds=[0,None], suffix=' us',
-                                            tip='Duration of pi pulse (RF) for ms=+1')         
-        self.treeDic_settings.add_parameter('Power_ms_-1', -20, 
-                                            type='float', step=1, 
-                                            bounds=[-50,30], suffix=' dBm',
-                                            tip='Constant power of the pi pulse of ms=-1')
-        self.treeDic_settings.add_parameter('Frequency_ms_-1', 3, 
-                                            type='float', step=0.1, 
-                                            bounds=[0,10], suffix=' GHz',
-                                            tip='Frequency of the pi pulse of ms=-1')     
-        self.treeDic_settings.add_parameter('dt_pi_pulse_ms_-1', 0.3, 
-                                            type='float', step=0.1, 
-                                            bounds=[0,None], suffix=' us',
-                                            tip='Duration of pi pulse (RF) for ms=-1')   
-
-        self.treeDic_settings.add_parameter('N_repeat_same_state', 3, 
-                                            type='int', step=1, 
-                                            bounds=[0,None],
-                                            tip='Number of time to repeat the measure of one state at the same time before measuring the next state.\nThis is useful if it is long for switching frequency.') 
-
-        self.treeDic_settings.add_parameter('PL0', 0.04, 
-                                            type='float', step=0.04, 
-                                            bounds=[0,None], 
-                                            tip='Mean photocounts from ms=0 for a SINGLE readout') 
-        self.treeDic_settings.add_parameter('Contrast', 0.1, 
-                                            type='float', step=0.01, 
-                                            bounds=[0,None], 
-                                            tip='Contrast between the state ms=0 and ms=+-1 at time=0.\nThis is defined as (PL0-PL+-)/PL0, where PL+- is the mean photocount of ms=+-1.') 
+                                            tip='Frequency of the RF') 
         
-        self.treeDic_settings.add_parameter('Rate_+_min', 0.01, 
+        self.treeDic_settings.add_parameter('t_probe', 10, 
                                             type='float', step=0.1, 
-                                            bounds=[0,None], suffix=' Hz',
-                                            tip='Guess for the minimum value of the rate gamma+') 
-        self.treeDic_settings.add_parameter('Rate_+_max', 150*1e3, 
+                                            bounds=[0,None], suffix=' us',
+                                            tip='Time to probe.') 
+
+        self.treeDic_settings.add_parameter('dt_pi_pulse_ms+1', 0.3, 
                                             type='float', step=0.1, 
-                                            bounds=[0,None], suffix=' Hz',
-                                            tip='Guess for the maximum value of the rate gamma+') 
-        self.treeDic_settings.add_parameter('Size_rate_+_axis', 150, 
-                                            type='int', step=10, 
-                                            bounds=[0,None],
-                                            tip='Number of points along the gamma+ axis for the pdf') 
-        self.treeDic_settings.add_parameter('Rate_-_min', 0.01, 
+                                            bounds=[0,None], suffix=' us',
+                                            tip='Duration of pi pulse for ms=+1(RF)') 
+        self.treeDic_settings.add_parameter('dt_pi_pulse_ms-1', 0.3, 
                                             type='float', step=0.1, 
-                                            bounds=[0,None], suffix=' Hz',
-                                            tip='Guess for the minimum value of the rate gamma-')         
-        self.treeDic_settings.add_parameter('Rate_-_max', 150*1e3, 
-                                            type='float', step=0.1, 
-                                            bounds=[0,None], suffix=' Hz',
-                                            tip='Guess for the maximum value of the rate gamma-')
-        self.treeDic_settings.add_parameter('Size_rate_-_axis', 150, 
-                                            type='int', step=10, 
-                                            bounds=[0,None],
-                                            tip='Number of points along the gamma- axis for the pdf') 
-        self.list_prior_types = ['Flat', 'Gaussian']
-        self.treeDic_settings.add_parameter('Prior_type', self.list_prior_types, 
-                                            tip='Which prior to use. Based on the bounds given for the rates.')    
+                                            bounds=[0,None], suffix=' us',
+                                            tip='Duration of pi pulse for ms=-1(RF)') 
         
         self.treeDic_settings.add_parameter('dt_laser_initiate', 3, 
                                             type='float', step=0.1, 
@@ -4041,35 +4002,17 @@ class GUIT1probeOneTime(egg.gui.Window):
         self.treeDic_settings.add_parameter('DIO_pulse_modulation_ms+1', 3, 
                                             type='int', step=1, 
                                             bounds=[0,16],
-                                            tip='DIO for modulating the pulse for ms+1. AKA for sending the RF.')  
-        self.treeDic_settings.add_parameter('DIO_pulse_modulation_ms-1', 3, 
+                                            tip='DIO for modulating the pi pulse of ms=+1. AKA for sending the RF.')  
+        self.treeDic_settings.add_parameter('DIO_pulse_modulation_ms-1', 4, 
                                             type='int', step=1, 
                                             bounds=[0,16],
-                                            tip='DIO for modulating the pulse for ms-1. AKA for sending the RF.')  
-        self.treeDic_settings.add_parameter('DIO_control_frequency', 4, 
-                                            type='int', step=1, 
-                                            bounds=[0,16],
-                                            tip='DIO for controlling which frequency is sent')  
+                                            tip='DIO for modulating the pi pulse of ms=-1. AKA for sending the RF.')  
+
         self.treeDic_settings.add_parameter('DIO_sync_scope', 5, 
                                             type='int', step=1, 
                                             bounds=[-1,16],
                                             tip='DIO for synchronizing the oscilloscope. Put -1 for nothing')
-        
-        
-        # Add the Image View for the posterior. 
-        self.plot_item = egg.pyqtgraph.PlotItem()
-        # Put an image in the plot item. 
-        self.plot_image = egg.pyqtgraph.ImageView(view=self.plot_item)
-        self.place_object(self.plot_image, row=2, column = 1, 
-                          row_span=2, column_span=3, alignment=0)  
-        self.set_column_stretch(1,10)
-        self.set_row_stretch(1,10)
-        
-        # Add a dummy map 
-        x = np.linspace(1, 20, 100)
-        y = np.linspace(1, 20, 100)
-        X,Y = np.meshgrid(x,y)
-        
+
         # Make a label for showing some estimate
         self.label_estimates = egg.gui.Label('We have the best T1 tracer gui on the market.')
         self.place_object(self.label_estimates, row=2, column=4)        
@@ -4081,12 +4024,11 @@ class GUIT1probeOneTime(egg.gui.Window):
             Prepare the signal generator
             Prepare the plot
         """
-        _debug('GUIAdaptiveT1Bayes: button_prepare_experiment_clicked')   
+        _debug('GUIT1probeOneTimes: button_prepare_experiment_clicked')   
 
         # Initiate the prior, functions, etc. 
         self.initiate_attributes()
         
-        self.compute_show_estimate()
         #Initialize the plot
         self.update_plot_posterior()
 
@@ -4096,103 +4038,14 @@ class GUIT1probeOneTime(egg.gui.Window):
         # Trigger a dummy function for signaling to prepare stuffs
         self.event_prepare_experiment()
         
-    def initiate_attributes(self):
-        """
-        Initiate the attribute from the parameter choosen. 
-        Prepare the prior based on the parameters choosen. 
-        """
-        _debug('GUIAdaptiveT1Bayes: initiate_attributes') 
-        
-        # Set the prior
-        # Extract the bounds
-        self.Gp_min       = self.treeDic_settings['Rate_+_min']   #Minimum guess for gamma plus (Hz) 
-        self.Gp_max       = self.treeDic_settings['Rate_+_max']  #Maximun guess for gamma plus (Hz)    
-        self.size_axis_Gp = self.treeDic_settings['Size_rate_+_axis'] # Size of the axis gamam plus
-        self.Gm_min       = self.treeDic_settings['Rate_-_min']   #Minimum guess for gamma minus (Hz) 
-        self.Gm_max       = self.treeDic_settings['Rate_-_max']  #Maximun guess for gamma minus (Hz)  
-        self.size_axis_Gm = self.treeDic_settings['Size_rate_-_axis'] # Size of the axis gamam minus
-        #Define the axis for the prior pdf      
-        self.gp_axis = np.linspace(self.Gp_min, self.Gp_max, self.size_axis_Gp) 
-        self.gm_axis = np.linspace(self.Gm_min,self. Gm_max, self.size_axis_Gm) 
-        # Set the prior according to the type of prior
-        if self.treeDic_settings['Prior_type'] == 'Flat':
-            #Define the prior 
-            self.prior = 1+np.zeros([len(self.gm_axis), len(self.gp_axis)])      
-            # No need to normalize it for now. 
-        if self.treeDic_settings['Prior_type'] == 'Gaussian':
-            # The priore will be gaussian, with the widtdt set by the bounds
-            X,Y = np.meshgrid(self.gp_axis,self.gm_axis)
-            x0 = np.mean(self.gp_axis)
-            y0 = np.mean(self.gm_axis)
-            dx = 0.5*(self.Gp_max-self.Gp_min) # Width in the gamma+ direction
-            dy = 0.5*(self.Gm_max-self.Gm_min) # Width in the gamma- direction
-            self.prior = np.exp(-( (X-x0)*(X-x0)/dx**2 + (Y-y0)*(Y-y0)/dy**2 )) # A non-skewed gaussian
-
-        # Initiate the class for processing the protocol
-        # Define the constants
-        self.PL0      = self.treeDic_settings['PL0']
-        self.contrast = self.treeDic_settings['Contrast']
-        constants = [self.PL0, self.contrast]
-        # Define the model functions that should describe our experiment. 
-        #TODO Have these models as an input and prepare the pulse sequence accordingly.
-        f0 = protocol_model.PLModel(constants).PL00
-        fp = protocol_model.PLModel(constants).PLp0
-        fm = protocol_model.PLModel(constants).PLm0
-        model_functions = [f0, fp, fm]
-            
-        # Initiate the class. This class will contain the posterior and everything else. 
-        self.protocol_processor = protocol_bayes.Bayes3Measure(model_functions, constants)
-        self.protocol_processor.initialize(self.gp_axis, self.gm_axis, self.prior)
-        
-        # Adjust the informaton
-        self.update_information()
-
-    def update_information(self):
-        """
-        Update the information from the object
-        """
-        _debug('GUIAdaptiveT1Bayes: update_information')
-        # Extract the information so far
-        self.t_probe   = self.protocol_processor.get_t_probe()
-        self.gp_axis   = self.protocol_processor.get_gp_axis()
-        self.gm_axis   = self.protocol_processor.get_gm_axis()
-        self.posterior = self.protocol_processor.get_posterior()
-        
-        #TODO Define "get" methods for that. Or just keep it as it is ;p
-        self.Gp_guess = self.protocol_processor.Gp_guess
-        self.Gm_guess = self.protocol_processor.Gm_guess
-        self.eGp_guess = self.protocol_processor.eGp_guess
-        self.eGm_guess = self.protocol_processor.eGm_guess
-        
-        
-    def update_plot_posterior(self):
-        """
-        Update the plot of the posterior. The title should be clear lol. 
-        """    
-        _debug('GUIAdaptiveT1Bayes: update_plot_posterior')
-        
- 
-        # Set the axis 
-        # Get the scale (AKA the spacing between two neighboor points on the image)
-        self.scale_x = (self.gp_axis.max()-self.gp_axis.min())/len(self.gp_axis)*1e-3
-        self.scale_y = (self.gm_axis.max()-self.gm_axis.min())/len(self.gm_axis)*1e-3
-        
-        self.plot_item.setLabel('bottom', text='Gamma + (kHz)')
-        self.plot_item.setLabel('left'  , text='Gamma - (kHz)')      
-        
-        # Set the image
-        self.plot_image.setImage(self.posterior.T,
-                                 pos=(self.gp_axis.min(), self.gm_axis.min()),
-                                 scale =(self.scale_x, self.scale_y) )
-        # magic method for the image to fill all the space
-        self.plot_image.view.setAspectLocked(False) # Input True for having the scaling right.         
+       
         
     def prepare_pulse_sequence(self):
         """
         Prepare the pulse sequence. 
         It generates the objet to be converted into a data array. 
         """
-        _debug('GUIAdaptiveT1Bayes: prepare_pulse_sequence')
+        _debug('GUIT1probeOneTimes: prepare_pulse_sequence')
         
 
         DIO_laser       = self.treeDic_settings['DIO_laser']
@@ -4395,48 +4248,7 @@ class GUIT1probeOneTime(egg.gui.Window):
         This is was should be done after that the pulse sequence is defined. 
         """        
         return         
-    
-    def compute_show_estimate(self):
-        """
-        Compute and show some estimate for the pulse sequence. 
-        """
-        _debug('GUIAdaptiveT1Bayes: compute_show_estimate')
-        
-        # See notebook of Michael Caouette-Mansour on July 24 2020 for details 
-        # of how the minimum time is estimated
-        
-        # Compute roughly the total time for a single sequence. 
-        # This is assuming that the dominating time are the times to probe
-        # Multiply by 3 because we do it three time, for each ms state
-        T_seq = 0
-        
-        C0 = self.PL0  # Mean count per each readout for ms=0
-        c  = self.contrast  # Contrast between the states
-        C1 = C0*(1-c) # Mean count per readout for ms=+-1
-        
-        # Compute the minimum number of readout for distinguishing the states
-        N_min = (np.sqrt(C0) + np.sqrt(C1))**2/(C0-C1)**2
-        
-        T_minimum = T_seq*N_min # Minimum elapsed time before distinguishing the states
-        
-        # It should be in us. Let's convert it in minutes
-        T_minutes = T_minimum*1e-6/60 
-        
-        # Set the text and the label
-        text =   'Assuming %0.3f count per readout on state ms=0.'%C0
-        text+= '\nAssuming %0.3f count per readout on state ms=+-1.'%C1
-        text+= '\nNote that the contrast is %0.1f percente'%c
-        text+='\nIt should take %d readout before distinguishing those states'%N_min
-        text+='\nWith the current sequence, this should take at least %0.2f minutes'%T_minutes
-        self.text_estimate = text  
-            
-        # Set the text and the label
-        text =   'Gamma_+ = %f +- %f kHz'%(self.Gp_guess*1e-3, self.eGp_guess*1e-3)
-        text+= '\nGamma_- = %f +- %f kHz'%(self.Gm_guess*1e-3, self.eGm_guess*1e-3)
-        text+= '\nT_probe = %f us'%(self.t_probe*1e6)
-        self.text_results = text
-        self.label_estimates.set_text(self.text_estimate +'\n\n' 
-                                      + self.text_results)      
+     
     
 if __name__ == '__main__':
     _fc._debug_enabled = False
